+++ src/Service/AppSync/src/Enum/DataSourceType.php
@@ -9,6 +9,7 @@ final class DataSourceType
 {
     public const AMAZON_DYNAMODB = 'AMAZON_DYNAMODB';
     public const AMAZON_ELASTICSEARCH = 'AMAZON_ELASTICSEARCH';
+    public const AMAZON_EVENTBRIDGE = 'AMAZON_EVENTBRIDGE';
     public const AMAZON_OPENSEARCH_SERVICE = 'AMAZON_OPENSEARCH_SERVICE';
     public const AWS_LAMBDA = 'AWS_LAMBDA';
     public const HTTP = 'HTTP';
@@ -20,6 +21,7 @@ final class DataSourceType
         return isset([
             self::AMAZON_DYNAMODB => true,
             self::AMAZON_ELASTICSEARCH => true,
+            self::AMAZON_EVENTBRIDGE => true,
             self::AMAZON_OPENSEARCH_SERVICE => true,
             self::AWS_LAMBDA => true,
             self::HTTP => true,